django/requirements.txt
Babel==2.7.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==2.2.6
django-extensions==2.2.5
django-grpc==0.1.8
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
Faker==2.0.2
Flask==1.1.1
Flask-Uploads==0.2.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
phonenumbers==8.10.21
pkg-resources==0.0.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
PyDrive==1.3.1
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
